# Conflicts:
#	core/src/main/c/windows/files.c
#	core/src/main/java/io/questdb/cairo/TableUtils.java
#	core/src/main/java/io/questdb/cairo/pool/WriterPool.java
#	core/src/main/java/io/questdb/cairo/wal/WalWriter.java
#	core/src/main/java/io/questdb/cairo/wal/seq/TableSequencerAPI.java
#	core/src/main/java/io/questdb/cairo/wal/seq/TableTransactionLog.java
#	core/src/main/java/io/questdb/log/LogAlertSocketWriter.java
#	core/src/main/java/io/questdb/std/Files.java
#	core/src/main/java/io/questdb/std/FilesFacade.java
#	core/src/main/java/io/questdb/std/FilesFacadeImpl.java
#	core/src/main/resources/io/questdb/bin/armlinux/libquestdb.so
#	core/src/main/resources/io/questdb/bin/armosx/libquestdb.dylib
#	core/src/main/resources/io/questdb/bin/freebsd/libquestdb.so
#	core/src/main/resources/io/questdb/bin/linux/libquestdb.so
#	core/src/main/resources/io/questdb/bin/osx/libquestdb.dylib
#	core/src/main/resources/io/questdb/bin/windows/libquestdb.dll
#	core/src/test/java/io/questdb/FilesTest.java
#	core/src/test/java/io/questdb/cairo/TableReaderTest.java
#	core/src/test/java/io/questdb/cairo/TableWriterTest.java
#	core/src/test/java/io/questdb/griffin/O3FailureTest.java

+++ core/src/main/c/freebsd/accept.c
@@ -27,6 +27,6 @@
 #include <sys/socket.h>
 
 JNIEXPORT jint JNICALL Java_io_questdb_network_Net_abortAccept
-        (JNIEnv *e, jclass cl, jlong fd) {
+        (JNIEnv *e, jclass cl, jint fd) {
     return shutdown((int) fd, SHUT_RDWR);
 }
