+++ core/src/main/java/io/questdb/cairo/ColumnPurgeOperator.java
@@ -47,7 +47,7 @@ public class ColumnPurgeOperator implements Closeable {
     private final int updateCompleteColumnWriterIndex;
     private long longBytes;
     private int pathTableLen;
-    private long purgeLogPartitionFd = -1L;
+    private int purgeLogPartitionFd = -1;
     private long purgeLogPartitionTimestamp = Long.MAX_VALUE;
     private TxReader txReader;
     private TxnScoreboard txnScoreboard;
@@ -156,7 +156,7 @@ public class ColumnPurgeOperator implements Closeable {
     private void closePurgeLogCompleteFile() {
         if (ff.closeChecked(purgeLogPartitionFd)) {
             LOG.info().$("closed purge log complete file [fd=").$(purgeLogPartitionFd).I$();
-            purgeLogPartitionFd = -1L;
+            purgeLogPartitionFd = -1;
         }
     }
 
@@ -307,7 +307,7 @@ public class ColumnPurgeOperator implements Closeable {
 
     private int readTableId(Path path) {
         final int INVALID_TABLE_ID = Integer.MIN_VALUE;
-        long fd = ff.openRO(path.trimTo(pathTableLen).concat(TableUtils.META_FILE_NAME).$());
+        int fd = ff.openRO(path.trimTo(pathTableLen).concat(TableUtils.META_FILE_NAME).$());
         if (fd < 0) {
             return INVALID_TABLE_ID;
         }
