+++ core/src/main/java/io/questdb/cairo/IDGenerator.java
@@ -34,7 +34,7 @@ public class IDGenerator implements Closeable {
     private final String uniqueIdFileName;
     private final long uniqueIdMemSize;
 
-    private long uniqueIdFd = -1;
+    private int uniqueIdFd = -1;
     private long uniqueIdMem = 0;
 
     public IDGenerator(CairoConfiguration configuration, String uniqueIdFileName) {
