+++ core/src/main/java/io/questdb/cairo/vm/MemoryCMARWImpl.java
@@ -35,7 +35,7 @@ import io.questdb.std.str.LPSZ;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-//contiguous mapped appendable readable writable
+// contiguous mapped appendable readable writable
 public class MemoryCMARWImpl extends AbstractMemoryCR implements MemoryCMARW, MemoryCARW, MemoryMAR {
     private static final Log LOG = LogFactory.getLog(MemoryCMARWImpl.class);
     private final Long256Acceptor long256Acceptor = this::putLong256;
@@ -129,7 +129,7 @@ public class MemoryCMARWImpl extends AbstractMemoryCR implements MemoryCMARW, Me
     }
 
     @Override
-    public long getFd() {
+    public int getFd() {
         return fd;
     }
 
@@ -155,7 +155,7 @@ public class MemoryCMARWImpl extends AbstractMemoryCR implements MemoryCMARW, Me
     }
 
     @Override
-    public void of(FilesFacade ff, long fd, @Nullable CharSequence name, long size, int memoryTag) {
+    public void of(FilesFacade ff, int fd, @Nullable CharSequence name, long size, int memoryTag) {
         close();
         assert fd > 0;
         this.ff = ff;
@@ -166,7 +166,7 @@ public class MemoryCMARWImpl extends AbstractMemoryCR implements MemoryCMARW, Me
     }
 
     @Override
-    public void of(FilesFacade ff, long fd, @Nullable CharSequence name, long extendSegmentSize, long size, int memoryTag) {
+    public void of(FilesFacade ff, int fd, @Nullable CharSequence name, long extendSegmentSize, long size, int memoryTag) {
         of(ff, fd, null, size, memoryTag);
         this.extendSegmentMsb = Numbers.msb(extendSegmentSize);
     }
@@ -188,7 +188,7 @@ public class MemoryCMARWImpl extends AbstractMemoryCR implements MemoryCMARW, Me
     }
 
     @Override
-    public void switchTo(long fd, long offset, byte truncateMode) {
+    public void switchTo(int fd, long offset, byte truncateMode) {
         close(true, truncateMode);
         this.fd = fd;
         map(ff, null, offset, memoryTag);
