+++ core/src/main/java/io/questdb/cutlass/line/udp/UdpLineChannel.java
@@ -32,7 +32,7 @@ import io.questdb.network.NetworkFacade;
 
 public final class UdpLineChannel implements LineChannel {
     private static final Log LOG = LogFactory.getLog(UdpLineChannel.class);
-    private final long fd;
+    private final int fd;
     private final NetworkFacade nf;
     private final long sockaddr;
