+++ core/src/test/java/io/questdb/cairo/TableReaderTest.java
@@ -1921,7 +1921,7 @@ public class TableReaderTest extends AbstractCairoTest {
                 temp.of(engine.getConfiguration().getRoot()).concat("dummy_non_existing_path").$();
                 ff = new FilesFacadeImpl() {
                     @Override
-                    public long openRO(LPSZ name) {
+                    public int openRO(LPSZ name) {
                         if (Chars.endsWith(name, TableUtils.META_FILE_NAME) && openCount.decrementAndGet() < 0) {
                             return Files.openRO(temp);
                         }
@@ -1957,10 +1957,10 @@ public class TableReaderTest extends AbstractCairoTest {
             try (Path temp = new Path()) {
                 temp.of(engine.getConfiguration().getRoot()).concat("dummy_non_existing_path").$();
                 ff = new FilesFacadeImpl() {
-                    long metaFd = -1;
+                    int metaFd = -1;
 
                     @Override
-                    public long length(long fd) {
+                    public long length(int fd) {
                         if (fd == metaFd) {
                             return Files.length(temp);
                         }
@@ -1976,7 +1976,7 @@ public class TableReaderTest extends AbstractCairoTest {
                     }
 
                     @Override
-                    public long openRO(LPSZ name) {
+                    public int openRO(LPSZ name) {
                         if (Chars.endsWith(name, TableUtils.META_FILE_NAME) && openCount.decrementAndGet() < 0) {
                             return metaFd = Files.openRO(name);
                         }
@@ -3918,20 +3918,20 @@ public class TableReaderTest extends AbstractCairoTest {
             boolean called = false;
 
             @Override
-            public boolean close(long fd) {
+            public boolean close(int fd) {
                 fds.remove(fd);
                 return super.close(fd);
             }
 
             @Override
-            public boolean closeChecked(long fd) {
+            public boolean closeChecked(int fd) {
                 fds.remove(fd);
                 return super.closeChecked(fd);
             }
 
             @Override
-            public long openRO(LPSZ name) {
-                long fd = super.openRO(name);
+            public int openRO(LPSZ name) {
+                int fd = super.openRO(name);
                 if (Chars.endsWith(name, dcol) || Chars.endsWith(name, icol)) {
                     fds.add(fd);
                     called = true;
@@ -4072,7 +4072,6 @@ public class TableReaderTest extends AbstractCairoTest {
 
             long blob = allocBlob();
             try {
-
                 // test if reader behaves correctly when table is empty
 
                 try (TableReader reader = new TableReader(configuration, "all")) {
@@ -4083,7 +4082,6 @@ public class TableReaderTest extends AbstractCairoTest {
                 }
 
                 try (TableReader reader = new TableReader(configuration, "all")) {
-
                     RecordCursor cursor = reader.getCursor();
                     // this combination of reload/iterate/reload is deliberate
                     // we make sure that reload() behavior is not affected by
